# openapi.yaml
openapi: 3.0.0
info:
  title: Project Spark Blueprint API
  version: v1
  description: API for Project Spark Blueprint application
servers:
  - url: http://localhost:5000/v1 # Assuming API v1 prefix is handled by Flask app
    description: Local development server
paths:
...
  /projects/:
    get:
      summary: List current user's projects
      operationId: listProjects
      tags:
        - Projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectOutput'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a new project
      operationId: createProject
      tags:
        - Projects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateInput'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOutput'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /projects/{project_id}:
    get:
      summary: Get a project by ID
      operationId: getProjectById
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOutput'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update a project by ID
      operationId: updateProject
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateInput'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOutput'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete a project by ID
      operationId: deleteProject
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Project deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /projects/{project_id}/members:
    get:
      summary: List members of a project
      operationId: listProjectMembers
      tags:
        - Project Members
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMemberOutput'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      summary: Add a member to a project
      operationId: addProjectMember
      tags:
        - Project Members
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMemberInput'
      responses:
        '200': # Service returns updated project, could also be 201 if member is a new sub-resource
          description: Member added successfully (returns updated project)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOutput' # Consistent with current route impl
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError' # If project not found
  /projects/{project_id}/members/{user_id}:
    delete:
      summary: Remove a member from a project
      operationId: removeProjectMember
      tags:
        - Project Members
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project
          schema:
            type: string
            format: uuid
        - name: user_id
          in: path
          required: true
          description: ID of the user to remove
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  schemas:
...
    ErrorResponse: # Generic error response
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    UserSimpleOutput: # Simplified user for nested representations
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
    ProjectStatusEnum:
      type: string
      enum:
        - planning
        - active
        - on-hold
        - completed
        - cancelled
      description: Status of the project
    ProjectPriorityEnum:
      type: string
      enum:
        - low
        - medium
        - high
      description: Priority of the project
    ProjectRoleEnum: # Role within a specific project
      type: string
      enum:
        - editor
        - viewer
      description: Role of a user within a project
    ProjectBase:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatusEnum'
        priority:
          $ref: '#/components/schemas/ProjectPriorityEnum'
        start_date:
          type: string
          format: date
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        parent_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the parent project, if any
        progress:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          default: 0
    ProjectCreateInput:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
      type: object
      required: # name is required from ProjectBase
        - name 
      properties:
        team_member_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
      # owner_id is set by the backend based on authenticated user
    ProjectUpdateInput:
      type: object # Not allOf ProjectBase, as all fields are optional for update
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatusEnum'
          nullable: true
        priority:
          $ref: '#/components/schemas/ProjectPriorityEnum'
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        parent_id:
          type: string
          format: uuid
          nullable: true
        progress:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          nullable: true
        team_member_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
    ProjectOutput:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        owner:
          $ref: '#/components/schemas/UserSimpleOutput'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserSimpleOutput'
          readOnly: true
    ProjectMemberOutput:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        role_in_project:
          $ref: '#/components/schemas/ProjectRoleEnum'
        user: # Details of the member user
          $ref: '#/components/schemas/UserSimpleOutput'
        added_at:
          type: string
          format: date-time
      required:
        - user_id
        - role_in_project
        - user
        - added_at
    AddProjectMemberInput:
      type: object
      required:
        - user_id
        - role_in_project
      properties:
        user_id:
          type: string
          format: uuid
          description: ID of the user to add to the project
        role_in_project:
          $ref: '#/components/schemas/ProjectRoleEnum'
          default: viewer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Adding common responses to avoid repetition
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: User does not have permissions to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequestError:
      description: The request payload is malformed or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: [] 
# The global security definition means all endpoints require bearerAuth by default.
# For public endpoints like /auth/login and /auth/register, 
# you need to override this at the operation level with `security: []`
# Let's adjust paths for /auth/login and /auth/register to explicitly have no security

# Adjusting auth paths to have no global security
# This should be done by finding these paths in the existing openapi.yaml and adding 'security: []'
# For this edit, I will assume the edit tool handles merging this correctly.
# If not, a separate edit focusing only on auth paths might be needed.

# Path adjustments for auth (example, needs to be merged with existing auth paths)
# /auth/register:
#   post:
#     security: [] 
#     # ... rest of the definition
# /auth/login:
#   post:
#     security: []
#     # ... rest of the definition
# /auth/refresh: # Refresh token might or might not need Bearer auth, depending on flow.
               # Assuming refresh token itself is the auth for this one, also no Bearer.
#   post:
#     security: [] 
#     # ... rest of the definition 
...